bug fix
不能检测用户非手册评估的bug 相关的数据模型

+++ WebService/WebSocket/MTMHealthCollectDTO.cs
@@ -11,12 +11,12 @@ namespace WebService
         public int eatingpreference { get; set; }
         public string smoking { get; set; }
         public string drinking { get; set; }
-        public int height { get; set; }
-        public int weight { get; set; }
-        public int waist { get; set; }
+        public double height { get; set; }
+        public double weight { get; set; }
+        public double waist { get; set; }
         public int heartrate { get; set; }
-        public int systolicpressure { get; set; }
-        public int diastolicpressure { get; set; }
+        public double systolicpressure { get; set; }
+        public double diastolicpressure { get; set; }
         public float fastsugar { get; set; }
         public float randomsugar { get; set; }
         public float HbA1c { get; set; }
